+++ src/main/java/com/intellij/plugins/haxe/buildsystem/lime/OpenflXmlFileDescription.java
@@ -2,10 +2,7 @@ package com.intellij.plugins.haxe.buildsystem.lime;
 
 import com.intellij.openapi.module.Module;
 import com.intellij.openapi.util.Iconable;
-import com.intellij.openapi.util.io.FileUtilRt;
-import com.intellij.psi.xml.XmlDocument;
 import com.intellij.psi.xml.XmlFile;
-import com.intellij.psi.xml.XmlTag;
 import icons.HaxeIcons;
 import lombok.NoArgsConstructor;
 import org.jetbrains.annotations.NotNull;
@@ -18,13 +15,20 @@ public class OpenflXmlFileDescription extends LimeXmlFileDescription {
 
 
     @Override
-    public Icon getFileIcon(@Iconable.IconFlags int flags) {
-        return HaxeIcons.OPENFL_LOGO;
+    public @Nullable Icon getFileIcon(@NotNull XmlFile file, @Iconable.IconFlags int flags) {
+        if(isOpenFlFile(file)) {
+            return HaxeIcons.OPENFL_LOGO;
+        }
+        return null;
     }
 
 
     @Override
     public boolean isMyFile(@NotNull XmlFile file, @Nullable Module module) {
+        return isOpenFlFile(file);
+    }
+
+    private boolean isOpenFlFile(@NotNull XmlFile file) {
         return LimeOpenFlUtil.isOpenFlFile(file);
     }