+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 ## 1.5.12
+* Changed: support for intellij 2024.3  & dropping support for 2023.x
+* Bugfix: Errors where not navigable  when OpenFL project file was outside module dir (Fixed by fstrace4)
 * Bugfix: Function types for inherited methods could in some cases show wrong signature when type parameters where used.
 * Bugfix: Auto-completion would in some cases show wrong types when type parameters where used. 
 * Bugfix: Fix NPE when evaluating callExpression made out of reifications